refactor(api): :art: remove backoffice module from api modules
