--- packages/arithmico-api/src/modules/backoffice/release/release.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ReleaseService } from './release.service';
-
-@Module({
-  providers: [ReleaseService],
-})
-export class ReleaseModule {}
