+++ packages/arithmico-api/src/modules/email/events/email-components/email-footer.tsx
