+++ packages/arithmico-api/src/modules/email/events/index.ts
