+++ packages/arithmico-api/src/modules/email/events/send-acitvation-email/send-activation-email.event.ts
